requirements-test.txt
mypy
pytest
pytest-asyncio
pytest-homeassistant-custom-component
PyYAML
